feat: RequirementModalAndDiscardAlert component
